Fix logging attack results summary to wandb
Logging Attack Results Summary to wandb is currently failing with
TypeError - Data row contained incompatible types.

The TypeError is because we were trying to log both Number and String
entries to the same column to wandb which is not allowed.

Fix involves converting string entries to float(Number) to maintain
same datatype throughout the column in the table.

Fixes #645