Fix WordCNN input embedding gradient calculation
get_grad function currently calculates gradients irrespective of
the inputs to the model. In models for which the inputs are not
transposed , the gradient is calculated incorrectly leading to errors
in functinos such as gradient_based word importance ranking used in
A2T recipe.

get_grad changed to correctly calculate grad with appropriate tensor
shape.

Fixes #601