Add check to match model output to dataset labels
During training, output nodes of models are not checked against the
number of labels present in the dataset. Leading to scenarios where if
the number of output nodes are lesser than labels present , we get a
"CUDA error: device-side assert triggered" during training.

Changes includes checks to compare output nodes with labels in dataset
and if labels in dataset are non-negative.