Fix adversarial dataset generation in trainer.py
As detailed in Issue #498 , the adversarial dataset generated in
_generate_adversarial_examples function in trainer.py was leading to
the column indicating the example type being masked/lost.
Due to the masking of the column all the examples were considered as
clean_examples which is not expected behaviour.

This masking of the column is due to the way the Dataset wrapper is
defined in textattack/datasets/dataset.py.

Commit includes changes to the data-structure of adversarial_examples
to suit the Dataset wrapper and changes to the condition statements in
collate_fn function in get_train_dataloader function.