Update dataset_args to correctly set dataset_split
In PR #533 , the third element of dataset_args needed to be set to args.dataset_split , but due to minor error the second element was being set as args.dataset_split instead.

This code lead to bug described in Issue #599. 

Commit includes appropriate fix to update the third element correctly.